Two patches for bzrobots contributed by Gilly, plus a fix for a broken nullpointer-check (if (!foo) -> if (foo))
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@16579 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzrobots/RCRobotPlayer.cxx
@@ -43,13 +43,13 @@ RCRobotPlayer::RCRobotPlayer(const PlayerId& _id,
 }
 
 
-void RCRobotPlayer::doUpdate(double dt)
+void RCRobotPlayer::doUpdate(float dt)
 {
   LocalPlayer::doUpdate(dt);
 }
 
 
-void RCRobotPlayer::doUpdateMotion(double dt)
+void RCRobotPlayer::doUpdateMotion(float dt)
 {
   if (isAlive()) {
     double timeNow = TimeKeeper::getCurrent().getSeconds();
